+++ bliss/CHANGELOG.md
@@ -1,10 +1,15 @@
-## [0.0.2] 2023-06-26
+## [0.0.2] 2023-06-28
 ### Added
 - Change log ;-)
+- settings copy on start finish
 
 ### Changes
 - Upgrade to bookworm
+- rework finish
+- remove redunant code
+- improve messages
+- config location
 
 ### Fixes
-- N/A
+- finish now uses bashio and not execlineb fixes systanx error